Fix build warnings on mips32

+++ tests/core/router/test_router_send_to_device.c
@@ -51,8 +51,8 @@ main(void)
 
     if (pkt_size != sizeof(payload)) {
         lgtd_errx(
-            1, "unexpected pkt size %d (expected %ld)",
-            pkt_size, sizeof(payload)
+            1, "unexpected pkt size %d (expected %ju)",
+            pkt_size, (uintmax_t)sizeof(payload)
         );
     }
